Feat: Actors features
If you pass Item.Eval in your actors messages we can use that do display how the item was evaluated.

In the case of lootnscoot I have also set up right click context to change the items Rule.

To do this with other scripts importing this. you will want to Pass the item.Eval rule and then go in and adjust the script detect and your command / function call to make the adjustments.